\name{safeguard.d}\alias{safeguard.d}\alias{safeguard.d2}\title{Safeguard power analysis for Cohen's d effect size and for Pearson's correlation}\description{These functions perform a safeguard power analysis starting from a Cohen’s d effect size (assuming a two samples t-test) or from a Pearson correlation coefficient.}\usage{safeguard.d(d, n.1, n.2, sig.level=.05, power=.8, conf=.80)safeguard.d2(m1, m2, pooledsd, n.1, n.2, sig.level=.05, power=.8, conf=.80)safeguard.r(r, n, sig.level=.05, power=.8, conf=.80)}\arguments{  \item{d}{Standardized mean difference, Cohen’s d}  \item{m1}{The mean value of the group 1}  \item{m2}{The mean value of the group 2}  \item{sd1}{The standard deviation of the group 1}  \item{sd1}{The standard deviation of the group 2}  \item{r}{Pearson’s correlation coefficient}  \item{n}{Sample size of the original study}  \item{n.1}{Sample size for group 1 of the original study}  \item{n.2}{Sample size for group 2 of the original study}   \item{sig.level}{Significance level (Type I error probability or alpha)}     \item{power}{Power of the test (1 minus Type II error probability or 1 minus beta)}     \item{conf}{Power of the test (1 minus Type II error probability or 1 minus beta)One-tailed confidence interval for computing the safeguard effect size, defaults to .80. It is required that .5<conf<1.}}   \details{Function safeguard.d computes the safeguard effect size (ds) as the lower limit of the specified confidence interval of d, subsequently a power analysis is performed using ds as input, as well as sig.level and power, to get the safeguard sample size. Function safeguard.d2 is a wrapper for safeguard.d that allows to specify the means and the standard deviations of the two groups instead of the Cohen’s d. Function safeguard.r takes as input the Pearson’s correlation coefficient instead of the Cohen’s d.}\value{A vector of three elements is returned. “Lower_d” (“Lower_r” in the output of safeguard.r) is the safeguard effect size, “N_required” is the safeguard sample size (ns) required for the replication study (in safeguard.d this is the overall sample size for the two groups) and “SSR” is the safeguard sample ratio of the original study, namely the ratio of the safeguard sample size over the original sample size, N_required /(n.1+n.2).}\references{Perugini, M., Gallucci, M., Costantini, G. (in preparation)Safeguard Power as a Guarantee Against Imprecise Power Estimates}\author{Giulio Costantini (giulio.costantini@unimib.it)}\examples{\dontrun{# most examples are taken from the main paperlibrary(pwr)library(MBESS)#safeguard power analysis for Cohen's d = .80 and n=50. Notice that safeguard.d#considers the one-tailed confidence interval, so we specify as input conf=.975 #and not .95safeguard.d(d=.8, n.1=25, n.2=25, sig.level=.05, power=.8, conf=0.975)#the same analysis step by step, including also the examples from the standard power analysis and the power analysis on the upper limit of the confidence interval. ci.smd(smd=.8, n.1=25, n.2=25, conf.level=.95)pw<-pwr.t.test(d=.8, sig.level=.05, power=.8, type="two.sample", alternative="greater")pw_lo<-pwr.t.test(d=.2192689, sig.level=.05, power=.8, type="two.sample", alternative="greater")pw_hi<-pwr.t.test(d=1.373031, sig.level=.05, power=.8, type="two.sample", alternative="greater")ceiling(pw_lo$n)*2ceiling(pw$n)*2ceiling(pw_hi$n)*2# the same analysis using safeguard.d2, i.e., specifying a value for the means#and for the pooled SD. We used means of 1.8 and 1, with a pooled SD of 1 to get#an effect size of d=.8safeguard.d2(m1=1.8, m2=1, pooledsd=1, n.1=25, n.2=25, conf=.975)#some hypothetical scenarios presented in the main paperpw_hi2<-pwr.t.test(d=1.1, sig.level=.05, power=.8, alternative="greater")ceiling(pw_hi2$n)*2pw_lo2<-pwr.t.test(d=.5, sig.level=.05, power=.8, alternative="greater")ceiling(pw_lo2$n)*2pwr.t.test(n=42/2, d=1.1, sig.level=.05, power=NULL, alternative="greater")pwr.t.test(n=42/2, d=.5, sig.level=.05, power=NULL, alternative="greater")# an example of usage of the function safeguard.rsafeguard.r(r=.4, n=100, sig.level=.05, power=.8, conf=.80)#The examples discussed in the text, about evaluating empirical evidence.safeguard.d(d=.5, n.1=30, n.2=30, conf=.8)safeguard.d(d=.5, n.1=90, n.2=90, conf=.8)safeguard.d(d=.8, n.1=100, n.2=100, conf=.8)safeguard.d(d=.8, n.1=100, n.2=100, conf=.95)#The examples in Table 2safeguard.d(d=.7, n.1=25, n.2=25)safeguard.d(d=.7, n.1=100, n.2=100)safeguard.d(d=1, n.1=15, n.2=15)safeguard.d(d=1.5, n.1=10, n.2=10)safeguard.d(d=.4, n.1=40, n.2=40)#what would happen if one used the sample sizes indicated by the standard#power analysis, in cases A and B of Table 2pwr.t.test(n=26, d=.45, type="two.sample", alternative="greater", power=NULL)pwr.t.test(n=26, d=.58, type="two.sample", alternative="greater", power=NULL)}}\keyword{qgraph}\keyword{sem}\keyword{cfa}\section{Notes}{If the safeguard effect size has the opposite sign of the original effect size the functions return a warning (this happens if the original effect size is significant at a certain p-level, one tailed, but conf is higher than one minus that p-level): should this happen, we suggest to lower the value of conf. Moreover in some situations it may happen that the value of the safeguard effect size is too high or too low to compute the exact value of the safeguard sample size. When the safeguard effect size is close to zero the N_required is set to Inf, while when it is much higher than zero the N_required is set to 2 (i.e., the minimum possible value for a sample size). In these situations, warning messages are returned.}